+++ programs/borrow-lending/src/endpoints/init_lending_market.rs
@@ -9,6 +9,8 @@ pub struct InitLendingMarket<'info> {
     #[account(signer)]
     pub owner: AccountInfo<'info>,
     pub admin_bot: AccountInfo<'info>,
+    #[account(executable)]
+    pub aldrin_amm: AccountInfo<'info>,
     #[account(zero)]
     pub lending_market: Account<'info, LendingMarket>,
 }
@@ -16,7 +18,8 @@ pub struct InitLendingMarket<'info> {
 pub fn handle(
     ctx: Context<InitLendingMarket>,
     currency: UniversalAssetCurrency,
-    compound_fee: PercentageInt,
+    leveraged_compound_fee: PercentageInt,
+    vault_compound_fee: PercentageInt,
     min_collateral_uac_value_for_leverage: SDecimal,
 ) -> ProgramResult {
     let accounts = ctx.accounts;
@@ -25,7 +28,9 @@ pub fn handle(
     accounts.lending_market.owner = accounts.owner.key();
     accounts.lending_market.currency = currency;
     accounts.lending_market.admin_bot = accounts.admin_bot.key();
-    accounts.lending_market.compound_fee = compound_fee;
+    accounts.lending_market.aldrin_amm = accounts.aldrin_amm.key();
+    accounts.lending_market.leveraged_compound_fee = leveraged_compound_fee;
+    accounts.lending_market.vault_compound_fee = vault_compound_fee;
     accounts
         .lending_market
         .min_collateral_uac_value_for_leverage =