+++ programs/borrow-lending/src/models/lending_market.rs
@@ -8,17 +8,27 @@ pub struct LendingMarket {
     /// Only a signer with this pubkey will be allowed to call the compound
     /// endpoint, the take reserve cap snapshot endpoint, etc.
     pub admin_bot: Pubkey,
-    /// How much of the collected reward is claimed by compound bot.
-    pub compound_fee: PercentageInt,
+    /// Whitelist Aldrin program ID for leverage yield farming (important) and
+    /// for vaults.
+    pub aldrin_amm: Pubkey,
+    /// How much of the collected reward is claimed by compound bot when
+    /// compounding leveraged position.
+    pub leveraged_compound_fee: PercentageInt,
+    /// How much of the collected reward is claimed on vault's compounding.
+    pub vault_compound_fee: PercentageInt,
     /// In order to open a leveraged position, the worth of collateral in the
     /// UAC must be at least this much.
     pub min_collateral_uac_value_for_leverage: SDecimal,
     pub currency: UniversalAssetCurrency,
+    /// Artificially pad some bytes for future configuration values. We use
+    /// `u64` instead of `u8` because std doesn't provide default
+    /// implementation for values larger than 32.
+    pub _padding: [u64; 16],
 }
 
 impl LendingMarket {
     pub fn space() -> usize {
-        131
+        292
     }
 }
 
@@ -31,7 +41,11 @@ mod tests {
         let lending_market = LendingMarket::default();
         assert_eq!(lending_market.owner, Pubkey::default());
         assert_eq!(lending_market.enable_flash_loans, false);
-        assert_eq!(lending_market.compound_fee, PercentageInt::new(0));
+        assert_eq!(
+            lending_market.leveraged_compound_fee,
+            PercentageInt::new(0)
+        );
+        assert_eq!(lending_market.vault_compound_fee, PercentageInt::new(0));
         assert_eq!(lending_market.currency, UniversalAssetCurrency::default());
         assert_eq!(lending_market.admin_bot, Pubkey::default());
         assert_eq!(