+++ tests/src/3-init-reserve.ts
@@ -100,6 +100,7 @@ export function test(
         program,
         owner,
         shmemProgramId,
+        undefined,
         Keypair.generate().publicKey
       );
       await expect(differentMarket.addReserve(10)).to.be.rejected;